0x00. Python - Variable Annotations Mandatory Task 6

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -5,7 +5,7 @@ Takes a list of mixed data types and returns a sum as float
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
     """
     Returns the sum of the ints and floats as a float
     """
